Cleaned code, RedditClient works so far, configured custom theme
The project now uses a function to configure the theme programmatically, will have to create a configuration file to get rid of the unecessary function. There is a RedditClient static variable now, created using lazy_static. I believe it was necessary to have the sufficient lifetime for the Cursive app. Some extraneous code I kept around for reference was also deleted, and remaining code was reformatted.